refactor: extract snapshot crate and remove anything else
As discussed with the Firecracker maintainers[1], we are forking
the snapshot crate part of Firecracker and maintain it separately, so
that we can publish it on crates.io and let fuse-backend-rs[2] and
dragonball[3] depend on it.

This is done by first forking the firecracker[4] repository and then
refactor the code directory to drop all the unrelated bits. Just so that
we can preseve as much the original git history as possible.

The initial code base is Firecracker release v1.5.0.

Kudos to the Firecracker community!

[1]: https://github.com/firecracker-microvm/firecracker/issues/4162
[2]: https://github.com/cloud-hypervisor/fuse-backend-rs
[3]: https://github.com/kata-containers/kata-containers/tree/main/src/dragonball
[4]: https://github.com/firecracker-microvm/firecracker

Signed-off-by: Peng Tao <bergwolf@hyper.sh>